refactor: change event-detail folder directory
