+++ src/components/event-detail/EventDetailItem.tsx
